refactor: widgets 레이어 SRP 정리 — EpigramForm 분리 + TabbedSection 해크 제거 + EpigramDetailCard 정리
계층별 순차 리팩토링 3단계. widgets 내부 오버크기 파일 분해, display:none UI 해크 제거, premature memo 제거, 외부 URL 검증.

EpigramForm 460 LOC 분리 (widgets/epigram-form/ui/fields/):
- ContentField — 내용 입력 + 글자수 카운터
- AuthorField — 저자 라디오 선택 + DIRECT 인풋 (AuthorRadioOption 로컬 유지)
- ReferenceFields — 출처 제목 + URL
- TagsField — 태그 입력 + 칩 (TagChip 로컬 유지, cross-cutting PR 에서 shared 이동 예정)
- FormActions — submit / cancel 버튼 조합
- FieldLabel — 공통 라벨
- FieldError — 공통 에러 표시
- EpigramForm.tsx — 메인 오케스트레이터만 남김 (~100 LOC)
getCounterClassName 은 컴포넌트 외부 pure 함수로 빼고 MAX_CONTENT_LENGTH 는 ContentField 에서 export 해서 메인이 재사용

TabbedSection:
- display:none 해크 제거: {visitedTabs.has(X) && <View style={{ display }}>} 패턴을 {activeTab === X ? <MyEpigramList /> : <MyCommentList />} 조건부 렌더링으로 교체. 비활성 탭 언마운트
- visitedTabs Set 상태 제거 — over-engineered 였고 각 리스트의 infinite query 가 이미 lazy 를 처리

EpigramDetailCard:
- RuledLines memo 제거 — props 없어 memo 이득 없음
- Reference 의 Linking.openURL 호출 전 URL 검증 추가. ALLOWED_URL_SCHEMES = [http, https] 로 허용 scheme 제한. 유효하지 않은 URL 은 링크 대신 plain text 로 표시

EpigramDetailList: 이미 shared 훅 (useScrollToFocusedInput) + memo 된 CommentItem 으로 정리되어 있어 추가 변경 없음.

tsc 통과.
